makedoc: when preparing the generated index, don't assume the index header exists. Instead, check for existance and assume the index is available - if not sphinx will bark anyway. This allows documentation to be built if no API docs should be generated

+++ doc/base.py
@@ -249,7 +249,10 @@ output: html"""
 		if self._sphinx:
 			self.remove_version_info('sphinx')
 			ip = self.index_rst_path()
-			if ip.isfile():
+			iph = ip+'.header'
+			# only remove index.rst if it appears we are generating it using 
+			# header and footer
+			if iph.isfile() and ip.isfile():
 				ip.remove()
 			# END remove generated index
 			
@@ -385,9 +388,14 @@ output: html"""
 		import mrv
 		
 		indexpath = self.index_rst_path()
+		index_header = indexpath+'.header'
+		if not index_header.isfile():
+			return
+		# END handle header doesn't exist
+		
 		ifp = open(indexpath, 'wb')
 		# write header
-		ifp.write((indexpath+'.header').bytes())
+		ifp.write(index_header.bytes())
 		
 		# write api index
 		if self._sphinx_autogen:
